@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
-apply plugin: 'com.neenbedankt.android-apt'
+apply plugin: 'me.tatarka.retrolambda'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.2"
 
     defaultConfig {
         applicationId "com.gb.canibuythat"
-        minSdkVersion 23
-        targetSdkVersion 23
+        minSdkVersion 21
+        targetSdkVersion 22
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -19,10 +19,14 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
-final ANDROID_VERSION = '23.3.0'
-final DAGGER_VERSION = '2.0.2'
+final ANDROID_VERSION = '25.3.0'
+final DAGGER_VERSION = '2.10'
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
@@ -50,11 +54,13 @@ dependencies {
 
     // Dagger 2
     compile "com.google.dagger:dagger:${DAGGER_VERSION}"
-    apt "com.google.dagger:dagger-compiler:${DAGGER_VERSION}"
+    annotationProcessor "com.google.dagger:dagger-compiler:${DAGGER_VERSION}"
 
     // Dagger 2 + test
-    testApt "com.google.dagger:dagger-compiler:${DAGGER_VERSION}"
+    testAnnotationProcessor "com.google.dagger:dagger-compiler:${DAGGER_VERSION}"
 
     // ButterKnife
     compile 'com.jakewharton:butterknife:6.1.0'
+
+    retrolambdaConfig "net.orfjackal.retrolambda:retrolambda:2.5.1"
 }
