@@ -1,7 +1,6 @@
 package com.gb.canibuythat.util;
 
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -13,23 +12,16 @@ public class ArrayUtils {
      * The items of the specified <code>array</code> are sorted based on the order in
      * which they occur in the specified <code>text</code>.
      */
-    public static String[] sortByOccurrence(String[] array, String text) {
+    static String[] sortByOccurrence(String[] array, String text) {
         final Map<String, Integer> indexMap = new HashMap<>();
-
-        for (int i = 0; i < array.length; i++) {
-            indexMap.put(array[i], text.indexOf(array[i]));
+        for (String anArray : array) {
+            indexMap.put(anArray, text.indexOf(anArray));
         }
-        Arrays.sort(array, new Comparator<String>() {
-            @Override public int compare(String lhs, String rhs) {
-                return indexMap.get(lhs)
-                        .compareTo(indexMap.get(rhs));
-            }
-        });
+        Arrays.sort(array, (lhs, rhs) -> indexMap.get(lhs).compareTo(indexMap.get(rhs)));
         return array;
     }
 
-    public static <T> String join(String delimiter, T[] tokens,
-            Stringifier<T> stringifier) {
+    public static <T> String join(String delimiter, T[] tokens, Stringifier<T> stringifier) {
         StringBuilder sb = new StringBuilder();
         boolean firstTime = true;
         int index = 0;
@@ -48,8 +40,7 @@ public class ArrayUtils {
         return sb.toString();
     }
 
-    public static <T> String join(String delimiter, Iterable<T> tokens,
-            Stringifier<T> stringifier) {
+    public static <T> String join(String delimiter, Iterable<T> tokens, Stringifier<T> stringifier) {
         StringBuilder sb = new StringBuilder();
         boolean firstTime = true;
         int index = 0;
@@ -65,6 +56,6 @@ public class ArrayUtils {
     }
 
     public interface Stringifier<T> {
-        public String toString(int index, T item);
+        String toString(int index, T item);
     }
 }
