@@ -4,7 +4,7 @@ import android.app.Activity;
 import android.app.Application;
 import android.os.Bundle;
 
-public class ActivityLifecycleCallbacksAdapter
+class ActivityLifecycleCallbacksAdapter
         implements Application.ActivityLifecycleCallbacks {
 
     @Override
