@@ -65,167 +65,154 @@ public class BudgetItemEditorFragment extends Fragment {
         DEFAULT_FIRST_OCCURRENCE_START = c.getTime();
     }
 
-    @InjectView(R.id.name) EditText mNameView;
-    @InjectView(R.id.amount) EditText mAmountView;
-    @InjectView(R.id.enabled) CheckBox mEnabledView;
-    @InjectView(R.id.category) Spinner mCategoryView;
-    @InjectView(R.id.first_occurence_start) Button mFirstOccurrenceStartView;
-    @InjectView(R.id.first_occurence_end) Button mFirstOccurrenceEndView;
-    @InjectView(R.id.occurence_count) EditText mOccurrenceLimitView;
-    @InjectView(R.id.period_multiplier) EditText mPeriodMultiplierView;
-    @InjectView(R.id.period_type) Spinner mPeriodTypeView;
-    @InjectView(R.id.notes) EditText mNotesView;
-    @InjectView(R.id.spending_events) TextView mSpendingEventsView;
+    @InjectView(R.id.name) EditText nameView;
+    @InjectView(R.id.amount) EditText amountView;
+    @InjectView(R.id.enabled) CheckBox enabledView;
+    @InjectView(R.id.category) Spinner categoryView;
+    @InjectView(R.id.first_occurence_start) Button firstOccurrenceStartView;
+    @InjectView(R.id.first_occurence_end) Button firstOccurrenceEndView;
+    @InjectView(R.id.occurence_count) EditText occurrenceLimitView;
+    @InjectView(R.id.period_multiplier) EditText periodMultiplierView;
+    @InjectView(R.id.period_type) Spinner periodTypeView;
+    @InjectView(R.id.notes) EditText notesView;
+    @InjectView(R.id.spending_events) TextView spendingEventsView;
 
-    private BudgetItem mOriginalBudgetItem;
-    private boolean mFirstOccurrenceStartDateChanged;
-    private boolean mFirstOccurrenceEndDateChanged;
+    private BudgetItem originalBudgetItem;
+    private boolean firstOccurrenceStartDateChanged;
+    private boolean firstOccurrenceEndDateChanged;
 
-    private DatePickerDialog mFirstOccurrenceStartPickerDialog;
-    private DatePickerDialog mFirstOccurrenceEndPickerDialog;
-    private MenuItem mDeleteButton;
-    private ViewGroup mRootView;
+    private DatePickerDialog firstOccurrenceStartPickerDialog;
+    private DatePickerDialog firstOccurrenceEndPickerDialog;
+    private MenuItem deleteButton;
+    private ViewGroup rootView;
     private View.OnTouchListener keyboardDismisser = new View.OnTouchListener() {
-
-        @Override public boolean onTouch(View v, MotionEvent event) {
-            if (event.getAction() == MotionEvent.ACTION_DOWN &&
-                    mRootView.getFocusedChild() != null) {
-                ViewUtils.hideKeyboard(mRootView.getFocusedChild());
+        @Override
+        public boolean onTouch(View v, MotionEvent event) {
+            if (event.getAction() == MotionEvent.ACTION_DOWN && rootView.getFocusedChild() != null) {
+                ViewUtils.hideKeyboard(rootView.getFocusedChild());
             }
             return false;
         }
     };
-    private DatePickerDialog.OnDateSetListener mDateSetListener =
-            new DatePickerDialog.OnDateSetListener() {
+    private DatePickerDialog.OnDateSetListener mDateSetListener = new DatePickerDialog.OnDateSetListener() {
 
-                @Override
-                public void onDateSet(DatePicker view, int year, int monthOfYear,
-                        int dayOfMonth) {
-                    Calendar c = Calendar.getInstance();
-                    c.set(Calendar.YEAR, year);
-                    c.set(Calendar.MONTH, monthOfYear);
-                    c.set(Calendar.DAY_OF_MONTH, dayOfMonth);
-                    DateUtils.clearLowerBits(c);
+        @Override
+        public void onDateSet(DatePicker view, int year, int monthOfYear, int dayOfMonth) {
+            Calendar c = Calendar.getInstance();
+            c.set(Calendar.YEAR, year);
+            c.set(Calendar.MONTH, monthOfYear);
+            c.set(Calendar.DAY_OF_MONTH, dayOfMonth);
+            DateUtils.clearLowerBits(c);
 
-                    switch ((int) view.getTag()) {
-                        case R.id.first_occurence_start:
-                            mFirstOccurrenceStartView.setText(
-                                    DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
-                            Date firstOccurrenceEnd = getFirstOccurrenceEndFromScreen();
+            switch ((int) view.getTag()) {
+                case R.id.first_occurence_start:
+                    firstOccurrenceStartView.setText(DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
+                    Date firstOccurrenceEnd = getFirstOccurrenceEndFromScreen();
 
-                            if (firstOccurrenceEnd.getTime() < c.getTime()
-                                    .getTime()) {
-                                mFirstOccurrenceEndPickerDialog =
-                                        new DatePickerDialog(getActivity(),
-                                                mDateSetListener, c.get(Calendar.YEAR),
-                                                c.get(Calendar.MONTH),
-                                                c.get(Calendar.DAY_OF_MONTH));
-                                applyFirstOccurrenceEndToScreen(c.getTime());
-                            }
-                            mFirstOccurrenceStartDateChanged = !c.getTime()
-                                    .equals(DEFAULT_FIRST_OCCURRENCE_START);
-                            break;
-                        case R.id.first_occurence_end:
-                            mFirstOccurrenceEndView.setText(
-                                    DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
-                            Date firstOccurrenceStart =
-                                    getFirstOccurrenceStartFromScreen();
-
-                            if (firstOccurrenceStart.getTime() > c.getTime()
-                                    .getTime()) {
-                                mFirstOccurrenceStartPickerDialog =
-                                        new DatePickerDialog(getActivity(),
-                                                mDateSetListener, c.get(Calendar.YEAR),
-                                                c.get(Calendar.MONTH),
-                                                c.get(Calendar.DAY_OF_MONTH));
-                                applyFirstOccurrenceStartToScreen(c.getTime());
-                            }
-                            mFirstOccurrenceEndDateChanged = !c.getTime()
-                                    .equals(DEFAULT_FIRST_OCCURRENCE_END);
-                            break;
+                    if (firstOccurrenceEnd.getTime() < c.getTime().getTime()) {
+                        firstOccurrenceEndPickerDialog = new DatePickerDialog(getActivity(),
+                                mDateSetListener, c.get(Calendar.YEAR),
+                                c.get(Calendar.MONTH),
+                                c.get(Calendar.DAY_OF_MONTH));
+                        applyFirstOccurrenceEndToScreen(c.getTime());
                     }
-                }
-            };
-    private View.OnClickListener mDatePickerOnClickListener = new View.OnClickListener() {
+                    firstOccurrenceStartDateChanged = !c.getTime().equals(DEFAULT_FIRST_OCCURRENCE_START);
+                    break;
+                case R.id.first_occurence_end:
+                    firstOccurrenceEndView.setText(DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
+                    Date firstOccurrenceStart = getFirstOccurrenceStartFromScreen();
 
-        @Override public void onClick(View v) {
+                    if (firstOccurrenceStart.getTime() > c.getTime().getTime()) {
+                        firstOccurrenceStartPickerDialog = new DatePickerDialog(getActivity(),
+                                mDateSetListener, c.get(Calendar.YEAR),
+                                c.get(Calendar.MONTH),
+                                c.get(Calendar.DAY_OF_MONTH));
+                        applyFirstOccurrenceStartToScreen(c.getTime());
+                    }
+                    firstOccurrenceEndDateChanged = !c.getTime().equals(DEFAULT_FIRST_OCCURRENCE_END);
+                    break;
+            }
+        }
+    };
+    private View.OnClickListener mDatePickerOnClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
             switch (v.getId()) {
                 case R.id.first_occurence_start:
                     getFirstOccurrenceStartPickerDialog().show();
-                    mFirstOccurrenceStartView.setError(null);
+                    firstOccurrenceStartView.setError(null);
                     break;
                 case R.id.first_occurence_end:
                     getFirstOccurrenceEndPickerDialog().show();
-                    mFirstOccurrenceEndView.setError(null);
+                    firstOccurrenceEndView.setError(null);
                     break;
             }
         }
     };
 
-    public BudgetItemEditorFragment() {
-    }
-
-    @Override public void onCreate(Bundle savedInstanceState) {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setHasOptionsMenu(true);
     }
 
-    @Override public View onCreateView(LayoutInflater inflater, ViewGroup container,
-            Bundle savedInstanceState) {
-        mRootView = (ViewGroup) inflater.inflate(R.layout.fragment_budget_item_editor,
-                container, false);
-        ButterKnife.inject(this, mRootView);
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        rootView = (ViewGroup) inflater.inflate(R.layout.fragment_budget_item_editor, container, false);
+        ButterKnife.inject(this, rootView);
 
-        mCategoryView.setAdapter(new PlusOneAdapter(BudgetItem.BudgetItemType.values()));
-        mCategoryView.setOnTouchListener(keyboardDismisser);
+        categoryView.setAdapter(new PlusOneAdapter(BudgetItem.BudgetItemType.values()));
+        categoryView.setOnTouchListener(keyboardDismisser);
 
-        mPeriodTypeView.setAdapter(new PlusOneAdapter(BudgetItem.PeriodType.values()));
-        mPeriodTypeView.setOnTouchListener(keyboardDismisser);
-        mPeriodTypeView.setOnItemSelectedListener(
-                new AdapterView.OnItemSelectedListener() {
+        periodTypeView.setAdapter(new PlusOneAdapter(BudgetItem.PeriodType.values()));
+        periodTypeView.setOnTouchListener(keyboardDismisser);
+        periodTypeView.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
 
-                    @Override public void onItemSelected(AdapterView<?> parent, View view,
-                            int position, long id) {
-                        mFirstOccurrenceEndView.setError(null);
-                    }
+            @Override
+            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+                firstOccurrenceEndView.setError(null);
+            }
 
-                    @Override public void onNothingSelected(AdapterView<?> parent) {
-                        // nothing to do
-                    }
-                });
+            @Override
+            public void onNothingSelected(AdapterView<?> parent) {
+                // nothing to do
+            }
+        });
 
-        mFirstOccurrenceStartView.setOnClickListener(mDatePickerOnClickListener);
-        mFirstOccurrenceStartView.setOnTouchListener(keyboardDismisser);
-        mFirstOccurrenceEndView.setOnClickListener(mDatePickerOnClickListener);
-        mFirstOccurrenceEndView.setOnTouchListener(keyboardDismisser);
+        firstOccurrenceStartView.setOnClickListener(mDatePickerOnClickListener);
+        firstOccurrenceStartView.setOnTouchListener(keyboardDismisser);
+        firstOccurrenceEndView.setOnClickListener(mDatePickerOnClickListener);
+        firstOccurrenceEndView.setOnTouchListener(keyboardDismisser);
 
         if (savedInstanceState != null) {
-            mOriginalBudgetItem = savedInstanceState.getParcelable(EXTRA_ITEM);
+            originalBudgetItem = savedInstanceState.getParcelable(EXTRA_ITEM);
         }
-        if (mOriginalBudgetItem == null && getArguments() != null &&
-                getArguments().containsKey(EXTRA_ITEM_ID)) {
+        if (originalBudgetItem == null && getArguments() != null && getArguments().containsKey(EXTRA_ITEM_ID)) {
             setContent(getArguments().getInt(EXTRA_ITEM_ID), true);
         } else {
             setContent(null, true);
         }
-        return mRootView;
+        return rootView;
     }
 
     private class PlusOneAdapter extends ArrayAdapter {
 
-        public PlusOneAdapter(Object[] items) {
+        PlusOneAdapter(Object[] items) {
             super(getActivity(), android.R.layout.simple_list_item_1, items);
         }
 
-        @Override public int getCount() {
+        @Override
+        public int getCount() {
             return super.getCount() + 1;
         }
 
-        @Override public Object getItem(int position) {
+        @Override
+        public Object getItem(int position) {
             return position == 0 ? "Select one" : super.getItem(position - 1);
         }
 
-        @Override public long getItemId(int position) {
+        @Override
+        public long getItemId(int position) {
             return position == 0 ? -1 : super.getItemId(position - 1);
         }
     }
@@ -239,37 +226,32 @@ public class BudgetItemEditorFragment extends Fragment {
     public void setContent(Integer budgetItemId, final boolean showKeyboardWhenDone) {
         if (budgetItemId != null) {
             new BudgetItemReadTask(budgetItemId, new Callback<BudgetItem>() {
-
-                @Override public void onSuccess(BudgetItem budgetItem) {
-                    BudgetItemEditorFragment.this.mOriginalBudgetItem = budgetItem;
-                    mFirstOccurrenceEndPickerDialog = null;
-                    mFirstOccurrenceStartPickerDialog = null;
+                @Override
+                public void onSuccess(BudgetItem budgetItem) {
+                    BudgetItemEditorFragment.this.originalBudgetItem = budgetItem;
+                    firstOccurrenceEndPickerDialog = null;
+                    firstOccurrenceStartPickerDialog = null;
                     applyBudgetItemToScreen(budgetItem);
 
                     if (showKeyboardWhenDone && isAdded()) {
-                        ViewUtils.showKeyboard(mNameView);
+                        ViewUtils.showKeyboard(nameView);
                     }
 
-                    if (mDeleteButton != null) {
-                        mDeleteButton.setVisible(true);
+                    if (deleteButton != null) {
+                        deleteButton.setVisible(true);
                     }
-                    Toast.makeText(App.getAppContext(), "BudgetItem loaded",
-                            Toast.LENGTH_SHORT)
-                            .show();
+                    Toast.makeText(App.getAppContext(), "BudgetItem loaded", Toast.LENGTH_SHORT).show();
                 }
 
-                @Override public void onFailure() {
-                    Toast.makeText(App.getAppContext(), "BudgetItem was not found",
-                            Toast.LENGTH_SHORT)
-                            .show();
+                @Override
+                public void onFailure() {
+                    Toast.makeText(App.getAppContext(), "BudgetItem was not found", Toast.LENGTH_SHORT).show();
                 }
 
-                @Override public void onError(Throwable t) {
+                @Override
+                public void onError(Throwable t) {
                     t.printStackTrace();
-                    Toast.makeText(App.getAppContext(),
-                            "Error loading data. Check logs for more information.",
-                            Toast.LENGTH_SHORT)
-                            .show();
+                    Toast.makeText(App.getAppContext(), "Error loading data. Check logs for more information.", Toast.LENGTH_SHORT).show();
                 }
             }).execute();
         } else {
@@ -277,45 +259,43 @@ public class BudgetItemEditorFragment extends Fragment {
         }
     }
 
-    @Override public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
         inflater.inflate(R.menu.edit, menu);
-        this.mDeleteButton = menu.findItem(R.id.menu_delete);
+        this.deleteButton = menu.findItem(R.id.menu_delete);
 
         if (getArguments() == null || !getArguments().containsKey(EXTRA_ITEM_ID)) {
-            mDeleteButton.setVisible(false);
+            deleteButton.setVisible(false);
         }
     }
 
-    @Override public boolean onOptionsItemSelected(MenuItem item) {
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.menu_save:
                 saveUserInputOrShowError();
                 break;
             case R.id.menu_delete:
-                if (mOriginalBudgetItem != null && mOriginalBudgetItem.isPersisted()) {
+                if (originalBudgetItem != null && originalBudgetItem.isPersisted()) {
                     new BudgetItemDeleteTask(new Callback<Boolean>() {
 
-                        @Override public void onSuccess(Boolean result) {
-                            mDeleteButton.setVisible(false);
-                            Toast.makeText(App.getAppContext(), "BudgetItem deleted",
-                                    Toast.LENGTH_SHORT)
-                                    .show();
+                        @Override
+                        public void onSuccess(Boolean result) {
+                            deleteButton.setVisible(false);
+                            Toast.makeText(App.getAppContext(), "BudgetItem deleted", Toast.LENGTH_SHORT).show();
                         }
 
-                        @Override public void onFailure() {
-                            Toast.makeText(App.getAppContext(),
-                                    "BudgetItem was not found", Toast.LENGTH_SHORT)
-                                    .show();
+                        @Override
+                        public void onFailure() {
+                            Toast.makeText(App.getAppContext(), "BudgetItem was not found", Toast.LENGTH_SHORT).show();
                         }
 
-                        @Override public void onError(Throwable t) {
+                        @Override
+                        public void onError(Throwable t) {
                             t.printStackTrace();
-                            Toast.makeText(App.getAppContext(),
-                                    "Error deleting data. Check logs for more " +
-                                            "information.", Toast.LENGTH_SHORT)
-                                    .show();
+                            Toast.makeText(App.getAppContext(), "Error deleting data. Check logs for more information.", Toast.LENGTH_SHORT).show();
                         }
-                    }, mOriginalBudgetItem.mId).execute();
+                    }, originalBudgetItem.id).execute();
                 }
                 break;
         }
@@ -324,14 +304,12 @@ public class BudgetItemEditorFragment extends Fragment {
 
     public void saveAndRun(Runnable onSaveOrDiscard) {
         if (shouldSave()) {
-            DialogUtils.getSaveOrDiscardDialog(getContext(),
-                    new DialogUtils.Executable() {
-
-                        @Override public boolean run() {
-                            return saveUserInputOrShowError();
-                        }
-                    }, onSaveOrDiscard)
-                    .show();
+            DialogUtils.getSaveOrDiscardDialog(getActivity(), new DialogUtils.Executable() {
+                @Override
+                public boolean run() {
+                    return saveUserInputOrShowError();
+                }
+            }, onSaveOrDiscard).show();
         } else {
             onSaveOrDiscard.run();
         }
@@ -345,39 +323,33 @@ public class BudgetItemEditorFragment extends Fragment {
         if (error == null) {
             final BudgetItem newBudgetItem = getBudgetItemFromScreen();
 
-            if (mOriginalBudgetItem != null) {
+            if (originalBudgetItem != null) {
                 // this will make an already saved item ot be updated instead of a new
                 // one being created
-                newBudgetItem.mId = mOriginalBudgetItem.mId;
-                newBudgetItem.mOrdering = mOriginalBudgetItem.mOrdering;
+                newBudgetItem.id = originalBudgetItem.id;
+                newBudgetItem.ordering = originalBudgetItem.ordering;
             }
-            new BudgetItemCreateOrUpdateTask(newBudgetItem,
-                    new Callback<Dao.CreateOrUpdateStatus>() {
+            new BudgetItemCreateOrUpdateTask(newBudgetItem, new Callback<Dao.CreateOrUpdateStatus>() {
 
-                        @Override public void onSuccess(Dao.CreateOrUpdateStatus result) {
-                            BudgetItemEditorFragment.this.mOriginalBudgetItem =
-                                    newBudgetItem;
-                            mDeleteButton.setVisible(true);
-                            loadSpendingOccurrences(newBudgetItem);
-                            Toast.makeText(App.getAppContext(), "BudgetItem saved",
-                                    Toast.LENGTH_SHORT)
-                                    .show();
-                        }
+                @Override
+                public void onSuccess(Dao.CreateOrUpdateStatus result) {
+                    BudgetItemEditorFragment.this.originalBudgetItem = newBudgetItem;
+                    deleteButton.setVisible(true);
+                    loadSpendingOccurrences(newBudgetItem);
+                    Toast.makeText(App.getAppContext(), "BudgetItem saved", Toast.LENGTH_SHORT).show();
+                }
 
-                        @Override public void onError(Throwable t) {
-                            if (t.getCause() != null && t.getCause()
-                                    .getCause() != null && t.getCause()
-                                    .getCause() instanceof SQLiteConstraintException) {
-                                t = t.getCause()
-                                        .getCause();
-                            }
-                            t.printStackTrace();
-                            Toast.makeText(App.getAppContext(),
-                                    "Error saving data. Check logs for more information.",
-                                    Toast.LENGTH_SHORT)
-                                    .show();
-                        }
-                    }).execute();
+                @Override
+                public void onError(Throwable t) {
+                    if (t.getCause() != null
+                            && t.getCause().getCause() != null
+                            && t.getCause().getCause() instanceof SQLiteConstraintException) {
+                        t = t.getCause().getCause();
+                    }
+                    t.printStackTrace();
+                    Toast.makeText(App.getAppContext(), "Error saving data. Check logs for more information.", Toast.LENGTH_SHORT).show();
+                }
+            }).execute();
             return true;
         } else {
             error.showError();
@@ -386,126 +358,105 @@ public class BudgetItemEditorFragment extends Fragment {
     }
 
     private void applyBudgetItemToScreen(final BudgetItem budgetItem) {
-        mNameView.setText(budgetItem.mName);
-
-        if (budgetItem.mAmount != null) {
-            mAmountView.setText(getString(R.string.detail_amount, budgetItem.mAmount));
+        nameView.setText(budgetItem.name);
+        if (budgetItem.amount != null) {
+            amountView.setText(getString(R.string.detail_amount, budgetItem.amount));
         } else {
-            mAmountView.setText(null);
+            amountView.setText(null);
         }
-        mEnabledView.setChecked(budgetItem.mEnabled);
-        if (budgetItem.mType != null) {
-            mCategoryView.setSelection(budgetItem.mType.ordinal() + 1);
+        enabledView.setChecked(budgetItem.enabled);
+        if (budgetItem.type != null) {
+            categoryView.setSelection(budgetItem.type.ordinal() + 1);
         }
 
-        Date firstOccurrenceStart = budgetItem.mFirstOccurrenceStart != null
-                                    ? budgetItem.mFirstOccurrenceStart
-                                    : DEFAULT_FIRST_OCCURRENCE_START;
+        Date firstOccurrenceStart = budgetItem.firstOccurrenceStart != null
+                ? budgetItem.firstOccurrenceStart
+                : DEFAULT_FIRST_OCCURRENCE_START;
         applyFirstOccurrenceStartToScreen(firstOccurrenceStart);
 
-        Date firstOccurrenceEnd =
-                budgetItem.mFirstOccurrenceEnd != null ? budgetItem.mFirstOccurrenceEnd
-                                                       : DEFAULT_FIRST_OCCURRENCE_END;
+        Date firstOccurrenceEnd = budgetItem.firstOccurrenceEnd != null ? budgetItem.firstOccurrenceEnd : DEFAULT_FIRST_OCCURRENCE_END;
         applyFirstOccurrenceEndToScreen(firstOccurrenceEnd);
 
-        if (budgetItem.mOccurrenceCount != null) {
-            mOccurrenceLimitView.setText(Integer.toString(budgetItem.mOccurrenceCount));
+        if (budgetItem.occurrenceCount != null) {
+            occurrenceLimitView.setText(Integer.toString(budgetItem.occurrenceCount));
         } else {
-            mOccurrenceLimitView.setText(null);
+            occurrenceLimitView.setText(null);
         }
 
-        if (budgetItem.mPeriodMultiplier != null) {
-            mPeriodMultiplierView.setText(Integer.toString(budgetItem.mPeriodMultiplier));
+        if (budgetItem.periodMultiplier != null) {
+            periodMultiplierView.setText(Integer.toString(budgetItem.periodMultiplier));
         } else {
-            mPeriodMultiplierView.setText(null);
+            periodMultiplierView.setText(null);
         }
 
-        if (budgetItem.mPeriodType != null) {
-            mPeriodTypeView.setSelection(budgetItem.mPeriodType.ordinal() + 1);
+        if (budgetItem.periodType != null) {
+            periodTypeView.setSelection(budgetItem.periodType.ordinal() + 1);
         }
-        mNotesView.setText(budgetItem.mNotes);
+        notesView.setText(budgetItem.notes);
         loadSpendingOccurrences(budgetItem);
     }
 
     private void clearScreen() {
-        mNameView.setText(null);
-        mAmountView.setText(null);
-        mEnabledView.setChecked(true);
-        mCategoryView.setSelection(0);
+        nameView.setText(null);
+        amountView.setText(null);
+        enabledView.setChecked(true);
+        categoryView.setSelection(0);
         applyFirstOccurrenceStartToScreen(DEFAULT_FIRST_OCCURRENCE_START);
         applyFirstOccurrenceEndToScreen(DEFAULT_FIRST_OCCURRENCE_END);
-        mOccurrenceLimitView.setText(null);
-        mPeriodMultiplierView.setText(null);
-        mPeriodTypeView.setSelection(0);
-        mNotesView.setText(null);
-        mSpendingEventsView.setText(null);
+        occurrenceLimitView.setText(null);
+        periodMultiplierView.setText(null);
+        periodTypeView.setSelection(0);
+        notesView.setText(null);
+        spendingEventsView.setText(null);
 
-        mOriginalBudgetItem = null;
-        mFirstOccurrenceEndPickerDialog = null;
-        mFirstOccurrenceStartPickerDialog = null;
+        originalBudgetItem = null;
+        firstOccurrenceEndPickerDialog = null;
+        firstOccurrenceStartPickerDialog = null;
     }
 
     private void loadSpendingOccurrences(final BudgetItem budgetItem) {
         BalanceReading balanceReading = UserPreferences.getBalanceReading();
-        BalanceCalculator.BalanceResult result = BalanceCalculator.get()
-                .getEstimatedBalance(budgetItem,
-                        balanceReading != null ? balanceReading.when
-                                                   : null,
-                        UserPreferences.getEstimateDate());
-        String spending = ArrayUtils.join("\n", result.spendingEvents,
-                new ArrayUtils.Stringifier<Date>() {
-
-                    @Override public String toString(int index, Date item) {
-                        return getString(R.string.spending_occurrence, index + 1,
-                                DateUtils.FORMAT_MONTH_DAY.format(item));
-                    }
-                });
+        BalanceCalculator.BalanceResult result = BalanceCalculator.get().getEstimatedBalance(budgetItem,
+                balanceReading != null ? balanceReading.when : null,
+                UserPreferences.getEstimateDate());
+        String spending = ArrayUtils.join("\n", result.spendingEvents, (index, item) -> getString(R.string.spending_occurrence, index + 1, DateUtils.FORMAT_MONTH_DAY.format(item)));
         spending = "Spent: " + result.bestCase + "/" + (result.worstCase) + "\n" + spending;
-        mSpendingEventsView.setText(spending);
+        spendingEventsView.setText(spending);
     }
 
     private void applyFirstOccurrenceStartToScreen(Date firstOccurrenceStart) {
-        mFirstOccurrenceStartView.setText(
-                DateUtils.FORMAT_MONTH_DAY.format(firstOccurrenceStart));
+        firstOccurrenceStartView.setText(DateUtils.FORMAT_MONTH_DAY.format(firstOccurrenceStart));
     }
 
     private void applyFirstOccurrenceEndToScreen(Date firstOccurrenceEnd) {
-        mFirstOccurrenceEndView.setText(
-                DateUtils.FORMAT_MONTH_DAY.format(firstOccurrenceEnd));
+        firstOccurrenceEndView.setText(DateUtils.FORMAT_MONTH_DAY.format(firstOccurrenceEnd));
     }
 
     private DatePickerDialog getFirstOccurrenceStartPickerDialog() {
         Calendar c = Calendar.getInstance();
-        if (mOriginalBudgetItem != null &&
-                mOriginalBudgetItem.mFirstOccurrenceStart != null) {
-            c.setTime(this.mOriginalBudgetItem.mFirstOccurrenceStart);
+        if (originalBudgetItem != null && originalBudgetItem.firstOccurrenceStart != null) {
+            c.setTime(this.originalBudgetItem.firstOccurrenceStart);
         }
-        if (mFirstOccurrenceStartPickerDialog == null) {
-            mFirstOccurrenceStartPickerDialog =
-                    new DatePickerDialog(getActivity(), mDateSetListener,
-                            c.get(Calendar.YEAR), c.get(Calendar.MONTH),
-                            c.get(Calendar.DAY_OF_MONTH));
-            mFirstOccurrenceStartPickerDialog.getDatePicker()
-                    .setTag(R.id.first_occurence_start);
+        if (firstOccurrenceStartPickerDialog == null) {
+            firstOccurrenceStartPickerDialog = new DatePickerDialog(getActivity(), mDateSetListener,
+                    c.get(Calendar.YEAR), c.get(Calendar.MONTH),
+                    c.get(Calendar.DAY_OF_MONTH));
+            firstOccurrenceStartPickerDialog.getDatePicker().setTag(R.id.first_occurence_start);
         }
-        return mFirstOccurrenceStartPickerDialog;
+        return firstOccurrenceStartPickerDialog;
     }
 
     private DatePickerDialog getFirstOccurrenceEndPickerDialog() {
         Calendar c = Calendar.getInstance();
-        if (mOriginalBudgetItem != null &&
-                mOriginalBudgetItem.mFirstOccurrenceEnd != null) {
-            c.setTime(this.mOriginalBudgetItem.mFirstOccurrenceEnd);
+        if (originalBudgetItem != null && originalBudgetItem.firstOccurrenceEnd != null) {
+            c.setTime(this.originalBudgetItem.firstOccurrenceEnd);
         }
-        if (mFirstOccurrenceEndPickerDialog == null) {
-            mFirstOccurrenceEndPickerDialog =
-                    new DatePickerDialog(getActivity(), mDateSetListener,
-                            c.get(Calendar.YEAR), c.get(Calendar.MONTH),
-                            c.get(Calendar.DAY_OF_MONTH));
-            mFirstOccurrenceEndPickerDialog.getDatePicker()
-                    .setTag(R.id.first_occurence_end);
+        if (firstOccurrenceEndPickerDialog == null) {
+            firstOccurrenceEndPickerDialog = new DatePickerDialog(getActivity(), mDateSetListener,
+                    c.get(Calendar.YEAR), c.get(Calendar.MONTH), c.get(Calendar.DAY_OF_MONTH));
+            firstOccurrenceEndPickerDialog.getDatePicker().setTag(R.id.first_occurence_end);
         }
-        return mFirstOccurrenceEndPickerDialog;
+        return firstOccurrenceEndPickerDialog;
     }
 
     /**
@@ -513,45 +464,38 @@ public class BudgetItemEditorFragment extends Fragment {
      * field (with {@link TextView#setError(CharSequence)} and the ones tha are shown
      * as a Toast or Dialog instead (DatePicker, etc...).
      */
-    public static class ValidationError {
-        public static final int TYPE_INPUT_FIELD = 1;
-        public static final int TYPE_NON_INPUT_FIELD = 2;
+    private static class ValidationError {
+        static final int TYPE_INPUT_FIELD = 1;
+        static final int TYPE_NON_INPUT_FIELD = 2;
 
-        public final int type;
-        public final View target;
-        public final String errorMessage;
+        final int type;
+        final View target;
+        final String errorMessage;
 
-        public ValidationError(int type, View target, String errorMessage) {
+        ValidationError(int type, View target, String errorMessage) {
             if (type == TYPE_INPUT_FIELD) {
                 if (target == null) {
-                    throw new IllegalArgumentException(
-                            "Please specify a target view for the input field error " +
-                                    "message");
+                    throw new IllegalArgumentException("Please specify a target view for the input field error " + "message");
                 }
                 if (!(target instanceof TextView)) {
-                    throw new IllegalArgumentException(
-                            "Wrong view type in ValidationError. Cannot show error " +
-                                    "message.");
+                    throw new IllegalArgumentException("Wrong view type in ValidationError. Cannot show error " + "message.");
 
                 }
             } else if (type != TYPE_NON_INPUT_FIELD) {
-                throw new IllegalArgumentException(
-                        "\"type\" must be one of {TYPE_INPUT_FIELD, " +
-                                "TYPE_NON_INPUT_FIELD}");
+                throw new IllegalArgumentException("\"type\" must be one of {TYPE_INPUT_FIELD, " + "TYPE_NON_INPUT_FIELD}");
             }
             this.type = type;
             this.target = target;
             this.errorMessage = errorMessage;
         }
 
-        public void showError() throws IllegalArgumentException {
+        void showError() throws IllegalArgumentException {
             if (type == TYPE_INPUT_FIELD) {
                 TextView textView = (TextView) target;
                 textView.setError(errorMessage);
                 textView.requestFocus();
             } else if (type == TYPE_NON_INPUT_FIELD) {
-                Toast.makeText(App.getAppContext(), errorMessage, Toast.LENGTH_SHORT)
-                        .show();
+                Toast.makeText(App.getAppContext(), errorMessage, Toast.LENGTH_SHORT).show();
                 if (target != null) {
                     target.requestFocus();
                 }
@@ -565,46 +509,37 @@ public class BudgetItemEditorFragment extends Fragment {
      * @return true if user input is valid
      */
     private ValidationError validateUserInput() {
-        if (TextUtils.isEmpty(mNameView.getText())) {
-            return new ValidationError(ValidationError.TYPE_INPUT_FIELD, mNameView,
-                    "Please specify a name");
+        if (TextUtils.isEmpty(nameView.getText())) {
+            return new ValidationError(ValidationError.TYPE_INPUT_FIELD, nameView, "Please specify a name");
         }
-        if (TextUtils.isEmpty(mAmountView.getText())) {
-            return new ValidationError(ValidationError.TYPE_INPUT_FIELD, mAmountView,
-                    "Please specify an amount");
+        if (TextUtils.isEmpty(amountView.getText())) {
+            return new ValidationError(ValidationError.TYPE_INPUT_FIELD, amountView, "Please specify an amount");
         }
-        if (!(mCategoryView.getSelectedItem() instanceof BudgetItem.BudgetItemType)) {
-            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null,
-                    "Please select a category");
+        if (!(categoryView.getSelectedItem() instanceof BudgetItem.BudgetItemType)) {
+            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null, "Please select a category");
         }
-        if (!(mPeriodTypeView.getSelectedItem() instanceof BudgetItem.PeriodType)) {
-            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null,
-                    "Please select a period");
+        if (!(periodTypeView.getSelectedItem() instanceof BudgetItem.PeriodType)) {
+            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null, "Please select a period");
         }
-        if (TextUtils.isEmpty(mPeriodMultiplierView.getText())) {
-            return new ValidationError(ValidationError.TYPE_INPUT_FIELD,
-                    mPeriodMultiplierView, "Please fill in");
+        if (TextUtils.isEmpty(periodMultiplierView.getText())) {
+            return new ValidationError(ValidationError.TYPE_INPUT_FIELD, periodMultiplierView, "Please fill in");
         }
         Date firstOccurrenceStart = getFirstOccurrenceStartFromScreen();
         Date firstOccurrenceEnd = getFirstOccurrenceEndFromScreen();
-
         if (firstOccurrenceStart.after(firstOccurrenceEnd)) {
-            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null,
-                    "Start date must not be higher then end date");
+            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, null, "Start date must not be higher then end date");
         }
 
         Calendar c = Calendar.getInstance();
         c.setTime(firstOccurrenceStart);
         DateUtils.clearLowerBits(c);
 
-        ((BudgetItem.PeriodType) mPeriodTypeView.getSelectedItem()).apply(c,
-                getPeriodMultiplierFromScreen());
+        ((BudgetItem.PeriodType) periodTypeView.getSelectedItem()).apply(c, getPeriodMultiplierFromScreen());
         c.add(Calendar.DAY_OF_MONTH, -1);
 
         if (firstOccurrenceEnd.after(c.getTime())) {
-            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD,
-                    mFirstOccurrenceEndView, "End date cannot be higher than " +
-                    DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
+            return new ValidationError(ValidationError.TYPE_NON_INPUT_FIELD, firstOccurrenceEndView,
+                    "End date cannot be higher than " + DateUtils.FORMAT_MONTH_DAY.format(c.getTime()));
         }
         return null;
     }
@@ -612,42 +547,36 @@ public class BudgetItemEditorFragment extends Fragment {
     private BudgetItem getBudgetItemFromScreen() {
         BudgetItem budgetItem = new BudgetItem();
         // title
-        if (!TextUtils.isEmpty(mNameView.getText())) {
-            budgetItem.mName = mNameView.getText()
-                    .toString();
+        if (!TextUtils.isEmpty(nameView.getText())) {
+            budgetItem.name = nameView.getText().toString();
         }
         // amount
-        if (!TextUtils.isEmpty(mAmountView.getText())) {
-            budgetItem.mAmount = Float.valueOf(mAmountView.getText()
-                    .toString());
+        if (!TextUtils.isEmpty(amountView.getText())) {
+            budgetItem.amount = Float.valueOf(amountView.getText().toString());
         }
         // enabled
-        budgetItem.mEnabled = mEnabledView.isChecked();
+        budgetItem.enabled = enabledView.isChecked();
         // type
-        if (mCategoryView.getSelectedItem() instanceof BudgetItem.BudgetItemType) {
-            budgetItem.mType =
-                    (BudgetItem.BudgetItemType) mCategoryView.getSelectedItem();
+        if (categoryView.getSelectedItem() instanceof BudgetItem.BudgetItemType) {
+            budgetItem.type = (BudgetItem.BudgetItemType) categoryView.getSelectedItem();
         }
         // firstOccurrenceStart
-        budgetItem.mFirstOccurrenceStart = getFirstOccurrenceStartFromScreen();
+        budgetItem.firstOccurrenceStart = getFirstOccurrenceStartFromScreen();
         // firstOccurrenceEnd
-        budgetItem.mFirstOccurrenceEnd = getFirstOccurrenceEndFromScreen();
+        budgetItem.firstOccurrenceEnd = getFirstOccurrenceEndFromScreen();
         // repetition
-        if (!TextUtils.isEmpty(mOccurrenceLimitView.getText())) {
-            budgetItem.mOccurrenceCount = Integer.valueOf(mOccurrenceLimitView.getText()
-                    .toString());
+        if (!TextUtils.isEmpty(occurrenceLimitView.getText())) {
+            budgetItem.occurrenceCount = Integer.valueOf(occurrenceLimitView.getText().toString());
         }
         // periodMultiplier
-        budgetItem.mPeriodMultiplier = getPeriodMultiplierFromScreen();
+        budgetItem.periodMultiplier = getPeriodMultiplierFromScreen();
         // period
-        if (mPeriodTypeView.getSelectedItem() instanceof BudgetItem.PeriodType) {
-            budgetItem.mPeriodType =
-                    (BudgetItem.PeriodType) mPeriodTypeView.getSelectedItem();
+        if (periodTypeView.getSelectedItem() instanceof BudgetItem.PeriodType) {
+            budgetItem.periodType = (BudgetItem.PeriodType) periodTypeView.getSelectedItem();
         }
         // notes
-        if (!TextUtils.isEmpty(mNotesView.getText())) {
-            budgetItem.mNotes = mNotesView.getText()
-                    .toString();
+        if (!TextUtils.isEmpty(notesView.getText())) {
+            budgetItem.notes = notesView.getText().toString();
         }
         return budgetItem;
     }
@@ -655,14 +584,12 @@ public class BudgetItemEditorFragment extends Fragment {
     private Date getFirstOccurrenceStartFromScreen() {
         Date firstOccurrenceDateStart;
 
-        if (mFirstOccurrenceStartPickerDialog != null) {
+        if (firstOccurrenceStartPickerDialog != null) {
             // user picked a date
-            firstOccurrenceDateStart = DateUtils.getDayFromDatePicker(
-                    mFirstOccurrenceStartPickerDialog.getDatePicker());
-        } else if (mOriginalBudgetItem != null &&
-                mOriginalBudgetItem.mFirstOccurrenceStart != null) {
+            firstOccurrenceDateStart = DateUtils.getDayFromDatePicker(firstOccurrenceStartPickerDialog.getDatePicker());
+        } else if (originalBudgetItem != null && originalBudgetItem.firstOccurrenceStart != null) {
             // user did not pick a date, but this is EDIT, not CREATE
-            firstOccurrenceDateStart = mOriginalBudgetItem.mFirstOccurrenceStart;
+            firstOccurrenceDateStart = originalBudgetItem.firstOccurrenceStart;
         } else {
             // user did not pick a date, and we are in CREATE mode
             firstOccurrenceDateStart = DEFAULT_FIRST_OCCURRENCE_START;
@@ -672,15 +599,12 @@ public class BudgetItemEditorFragment extends Fragment {
 
     private Date getFirstOccurrenceEndFromScreen() {
         Date firstOccurrenceDateEnd;
-
-        if (mFirstOccurrenceEndPickerDialog != null) {
+        if (firstOccurrenceEndPickerDialog != null) {
             // user picked a date
-            firstOccurrenceDateEnd = DateUtils.getDayFromDatePicker(
-                    mFirstOccurrenceEndPickerDialog.getDatePicker());
-        } else if (mOriginalBudgetItem != null &&
-                mOriginalBudgetItem.mFirstOccurrenceEnd != null) {
+            firstOccurrenceDateEnd = DateUtils.getDayFromDatePicker(firstOccurrenceEndPickerDialog.getDatePicker());
+        } else if (originalBudgetItem != null && originalBudgetItem.firstOccurrenceEnd != null) {
             // user did not pick a date, but this is EDIT, not CREATE
-            firstOccurrenceDateEnd = mOriginalBudgetItem.mFirstOccurrenceEnd;
+            firstOccurrenceDateEnd = originalBudgetItem.firstOccurrenceEnd;
         } else {
             // user did not pick a date, and we are in CREATE mode
             firstOccurrenceDateEnd = DEFAULT_FIRST_OCCURRENCE_END;
@@ -690,15 +614,12 @@ public class BudgetItemEditorFragment extends Fragment {
 
     private Integer getPeriodMultiplierFromScreen() {
         Integer periodMultiplier = null;
-
-        if (!TextUtils.isEmpty(mPeriodMultiplierView.getText())) {
+        if (!TextUtils.isEmpty(periodMultiplierView.getText())) {
             // user entered value
-            periodMultiplier = Integer.valueOf(mPeriodMultiplierView.getText()
-                    .toString());
-        } else if (mOriginalBudgetItem != null &&
-                mOriginalBudgetItem.mPeriodMultiplier != null) {
+            periodMultiplier = Integer.valueOf(periodMultiplierView.getText().toString());
+        } else if (originalBudgetItem != null && originalBudgetItem.periodMultiplier != null) {
             // user did not enter a value, but this is EDIT, not CREATE
-            periodMultiplier = mOriginalBudgetItem.mPeriodMultiplier;
+            periodMultiplier = originalBudgetItem.periodMultiplier;
         }
         return periodMultiplier;
     }
@@ -709,12 +630,9 @@ public class BudgetItemEditorFragment extends Fragment {
      */
     private boolean shouldSave() {
         BudgetItem userInput = getBudgetItemFromScreen();
-        boolean isNew = mOriginalBudgetItem == null &&
-                !new BudgetItem().compareForEditing(userInput,
-                        !(mFirstOccurrenceStartDateChanged ||
-                                mFirstOccurrenceEndDateChanged));
-        boolean changed = mOriginalBudgetItem != null &&
-                !mOriginalBudgetItem.compareForEditing(userInput, false);
+        boolean isNew = originalBudgetItem == null
+                && !new BudgetItem().compareForEditing(userInput, !(firstOccurrenceStartDateChanged || firstOccurrenceEndDateChanged));
+        boolean changed = originalBudgetItem != null && !originalBudgetItem.compareForEditing(userInput, false);
         return isNew || changed;
     }
 }
