@@ -34,9 +34,7 @@ import com.gb.canibuythat.ui.task.budget_item.MoveToIndexTask;
  * Activities containing this fragment MUST implement the {@link FragmentCallback}
  * interface.
  */
-public class BudgetItemListFragment extends Fragment
-        implements LoaderManager.LoaderCallbacks<Cursor>, AdapterView.OnItemClickListener,
-        DragNDropListView.OnItemDragNDropListener {
+public class BudgetItemListFragment extends Fragment implements LoaderManager.LoaderCallbacks<Cursor>, AdapterView.OnItemClickListener, DragNDropListView.OnItemDragNDropListener {
 
     /**
      * A dummy implementation of the {@link FragmentCallback} interface that does
@@ -44,18 +42,14 @@ public class BudgetItemListFragment extends Fragment
      * only when this fragment is not
      * attached to an activity.
      */
-    private static FragmentCallback dummyFragmentCallback = new FragmentCallback() {
-
-        @Override public void onListItemClick(int id) {
-        }
-    };
+    private static FragmentCallback dummyFragmentCallback = id -> {};
     /**
      * The fragment's current callback object, which is notified of list item clicks.
      */
-    private FragmentCallback mCallback = dummyFragmentCallback;
+    private FragmentCallback callback = dummyFragmentCallback;
 
-    private DragNDropListView mList;
-    private DragNDropCursorAdapter mAdapter;
+    private DragNDropListView list;
+    private DragNDropCursorAdapter adapter;
 
     /**
      * Mandatory empty constructor for the fragment manager to instantiate the fragment
@@ -64,98 +58,96 @@ public class BudgetItemListFragment extends Fragment
     public BudgetItemListFragment() {
     }
 
-    @Nullable @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
-            @Nullable Bundle savedInstanceState) {
-        View root =
-                inflater.inflate(R.layout.fragment_budget_item_list, container, false);
-        mList = (DragNDropListView) root.findViewById(android.R.id.list);
-        mList.setOnItemDragNDropListener(this);
-        mList.setOnItemClickListener(this);
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        View root = inflater.inflate(R.layout.fragment_budget_item_list, container, false);
+        list = (DragNDropListView) root.findViewById(android.R.id.list);
+        list.setOnItemDragNDropListener(this);
+        list.setOnItemClickListener(this);
         return root;
     }
 
-    @Override public void onViewCreated(View view, Bundle savedInstanceState) {
+    @Override
+    public void onViewCreated(View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         getLoaderManager().initLoader(hashCode(), null, this);
     }
 
     private DragNDropCursorAdapter getListAdapter() {
-        if (mAdapter == null) {
-            mAdapter = new BudgetItemListAdapter(getActivity(), null);
+        if (adapter == null) {
+            adapter = new BudgetItemListAdapter(getActivity(), null);
 
-            if (mList != null) {
-                mList.setDragNDropAdapter(mAdapter);
+            if (list != null) {
+                list.setDragNDropAdapter(adapter);
             }
         }
-        return mAdapter;
+        return adapter;
     }
 
-    @Override public void onAttach(Context context) {
+    @Override
+    public void onAttach(Context context) {
         super.onAttach(context);
 
         // Activities containing this fragment must implement its callback.
         if (!(getActivity() instanceof FragmentCallback)) {
-            throw new IllegalStateException(
-                    "Activity must implement fragment's callback.");
+            throw new IllegalStateException("Activity must implement fragment's callback.");
         }
 
-        mCallback = (FragmentCallback) getActivity();
+        callback = (FragmentCallback) getActivity();
     }
 
 
-    @Override public Loader<Cursor> onCreateLoader(int id, Bundle args) {
-        return new CursorLoader(getActivity(), BudgetProvider.BUDGET_ITEMS_URI, null,
-                null, null, Contract.BudgetItem.ORDERING + " ASC");
+    @Override
+    public Loader<Cursor> onCreateLoader(int id, Bundle args) {
+        return new CursorLoader(getActivity(), BudgetProvider.BUDGET_ITEMS_URI, null, null, null, Contract.BudgetItem.ORDERING + " ASC");
     }
 
-    @Override public void onLoadFinished(Loader<Cursor> loader, Cursor data) {
+    @Override
+    public void onLoadFinished(Loader<Cursor> loader, Cursor data) {
         getListAdapter().swapCursor(data);
     }
 
-    @Override public void onLoaderReset(Loader<Cursor> loader) {
+    @Override
+    public void onLoaderReset(Loader<Cursor> loader) {
         getListAdapter().swapCursor(null);
     }
 
-    @Override public void onDetach() {
+    @Override
+    public void onDetach() {
         super.onDetach();
 
         // Reset the active callback interface to the dummy implementation.
-        mCallback = dummyFragmentCallback;
+        callback = dummyFragmentCallback;
     }
 
     @Override
     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
         // Notify the active callback interface (the activity, if the
         // fragment is attached to one) that an item has been selected.
-        mCallback.onListItemClick((int) getListAdapter().getItemId(position));
+        callback.onListItemClick((int) getListAdapter().getItemId(position));
     }
 
     @Override
-    public void onItemDragStart(DragNDropListView parent, View view, int position,
-            long id) {
+    public void onItemDragStart(DragNDropListView parent, View view, int position, long id) {
 
     }
 
     @Override
-    public void onItemDrop(DragNDropListView parent, View view, final int startPosition,
-            final int endPosition, long id) {
+    public void onItemDrop(DragNDropListView parent, View view, final int startPosition, final int endPosition, long id) {
         new MoveToIndexTask(startPosition, endPosition, mReorderCallback).execute();
     }
 
     private Callback<Integer> mReorderCallback = new Callback<Integer>() {
-        @Override public void onError(Throwable t) {
-            if (t.getCause() != null && t.getCause()
-                    .getCause() != null && t.getCause()
-                    .getCause() instanceof SQLiteConstraintException) {
-                t = t.getCause()
-                        .getCause();
+        @Override
+        public void onError(Throwable t) {
+            if (t.getCause() != null
+                    && t.getCause().getCause() != null
+                    && t.getCause().getCause() instanceof SQLiteConstraintException) {
+                t = t.getCause().getCause();
             }
             t.printStackTrace();
-            Toast.makeText(App.getAppContext(),
-                    "Error saving data. Check logs for more information.",
-                    Toast.LENGTH_SHORT)
-                    .show();
+            Toast.makeText(App.getAppContext(), "Error saving data. Check logs for more information.", Toast.LENGTH_SHORT).show();
         }
     };
 
@@ -163,7 +155,7 @@ public class BudgetItemListFragment extends Fragment
      * A callback interface that all activities containing this fragment must implement.
      * This mechanism allows activities to be notified of events in the fragment.
      */
-    public interface FragmentCallback {
+    interface FragmentCallback {
 
         /**
          * Callback for when a budget item has been selected.
