@@ -3,35 +3,36 @@ package com.gb.canibuythat.ui.task;
 import android.os.AsyncTask;
 
 public abstract class TaskBase<R> extends AsyncTask<Void, Void, R> {
-
-    private Callback<R> mCallback;
-    private Throwable mError;
+    private Callback<R> callback;
+    private Throwable error;
 
     public TaskBase(Callback<R> callback) {
-        mCallback = callback;
+        this.callback = callback;
     }
 
-    @Override protected R doInBackground(Void... params) {
+    @Override
+    protected R doInBackground(Void... params) {
         try {
             return doWork();
         } catch (Throwable e) {
             e.printStackTrace();
-            mError = e;
+            error = e;
         }
         return null;
     }
 
     protected abstract R doWork() throws Exception;
 
-    @Override protected void onPostExecute(R result) {
-        if (mCallback != null) {
-            if (mError != null) {
-                mCallback.onError(mError);
+    @Override
+    protected void onPostExecute(R result) {
+        if (callback != null) {
+            if (error != null) {
+                callback.onError(error);
             } else {
                 if (result != null) {
-                    mCallback.onSuccess(result);
+                    callback.onSuccess(result);
                 } else {
-                    mCallback.onFailure();
+                    callback.onFailure();
                 }
             }
         }
