@@ -2,7 +2,7 @@ package com.gb.canibuythat.util;
 
 public class DirReadException extends Exception {
 
-    public DirReadException(String detailMessage) {
+    DirReadException(String detailMessage) {
         super(detailMessage);
     }
 }
