@@ -16,7 +16,8 @@ import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.util.Map;
 
-@SuppressWarnings({"SameParameterValue", "unused"}) public class PrefsUtil {
+@SuppressWarnings({"SameParameterValue", "unused"})
+public class PrefsUtil {
 
     private static final String PREFS_NAME = "settings";
     private static final String SEPARATOR = "dfg,hsdfk__jg34n95t";
@@ -55,12 +56,9 @@ import java.util.Map;
         }
         try {
             byte[] bytes = Base64.decode(data, 0);
-            ObjectInputStream ois = new ObjectInputStream(
-                    new ByteArrayInputStream(bytes, 0, bytes.length));
+            ObjectInputStream ois = new ObjectInputStream(new ByteArrayInputStream(bytes, 0, bytes.length));
             return (Map) ois.readObject();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (ClassNotFoundException e) {
+        } catch (IOException | ClassNotFoundException e) {
             e.printStackTrace();
         }
         return null;
@@ -69,32 +67,31 @@ import java.util.Map;
 
     public static void put(String key, Map map) {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
-
+        ObjectOutputStream oos = null;
         try {
-            ObjectOutputStream oos;
             oos = new ObjectOutputStream(baos);
             oos.writeObject(map);
             oos.flush();
             put(key, Base64.encodeToString(baos.toByteArray(), 0));
-            if (oos != null) {
-                oos.close();
-            }
         } catch (IOException e) {
             e.printStackTrace();
+        } finally {
+            try {
+                oos.close();
+            } catch (Throwable e) {
+                e.printStackTrace();
+            }
         }
     }
 
-
     public static String[] get(String key, String[] defaultValues) {
         String defaultValuesStr;
-
         if (defaultValues == null) {
             defaultValuesStr = null;
         } else {
             defaultValuesStr = TextUtils.join(SEPARATOR, defaultValues);
         }
         String text = get(key, defaultValuesStr);
-
         if (!TextUtils.isEmpty(text)) {
             return text.split(SEPARATOR);
         } else {
@@ -102,99 +99,77 @@ import java.util.Map;
         }
     }
 
-
     public static void put(String key, String[] values) {
         put(key, TextUtils.join(SEPARATOR, values));
     }
-
 
     public static void put(String key, boolean value) {
         getSecurePreferences().put(key, Boolean.toString(value));
     }
 
-
     public static boolean get(String key, boolean defaultValue) {
         String value = getSecurePreferences().getString(key);
-        return TextUtils.isEmpty(value) ? defaultValue : Boolean.valueOf(value)
-                .booleanValue();
+        return TextUtils.isEmpty(value) ? defaultValue : Boolean.valueOf(value);
     }
-
 
     public static void put(String key, String value) {
         getSecurePreferences().put(key, value);
     }
-
 
     public static String get(String key, String defaultValue) {
         String value = getSecurePreferences().getString(key);
         return TextUtils.isEmpty(value) ? defaultValue : value;
     }
 
-
     public static void put(String key, int value) {
         getSecurePreferences().put(key, Integer.toString(value));
     }
 
-
     public static int get(String key, int defaultValue) {
         String value = getSecurePreferences().getString(key);
-        return TextUtils.isEmpty(value) ? defaultValue : Integer.valueOf(value)
-                .intValue();
+        return TextUtils.isEmpty(value) ? defaultValue : Integer.valueOf(value);
     }
-
 
     public static void put(String key, long value) {
         getSecurePreferences().put(key, Long.toString(value));
     }
 
-
     public static long get(String key, long defaultValue) {
         String value = getSecurePreferences().getString(key);
-        return TextUtils.isEmpty(value) ? defaultValue : Long.valueOf(value)
-                .longValue();
+        return TextUtils.isEmpty(value) ? defaultValue : Long.valueOf(value);
     }
-
 
     public static void put(String key, float value) {
         getSecurePreferences().put(key, Float.toString(value));
     }
 
-
     public static float get(String key, float defaultValue) {
         String value = getSecurePreferences().getString(key);
-        return TextUtils.isEmpty(value) ? defaultValue : Float.valueOf(value)
-                .floatValue();
+        return TextUtils.isEmpty(value) ? defaultValue : Float.valueOf(value);
     }
-
 
     public static void remove(String key) {
         getSecurePreferences().removeValue(key);
     }
 
-
     private static SecurePreferences getSecurePreferences() {
         if (securePreferences == null) {
-            securePreferences = new SecurePreferences(App.getAppContext(), PREFS_NAME,
-                    App.generateUDID(), true);
+            securePreferences = new SecurePreferences(App.getAppContext(), PREFS_NAME, App.generateUDID(), true);
         }
         return securePreferences;
     }
 
-
     /**
-     * Registers a callback to be invoked when a change happens to the specified
-     * preference.
+     * Registers a callback to be invoked when a change happens to the specified preference.
      *
      * @param key      Preference key for which the specified callback should be
-     *                    registered to
+     *                 registered to
      * @param listener The callback that will run.
      * @see #unregisterOnSharedPreferenceChangeListener
      */
-    public static void registerOnSharedPreferenceChangeListener(String key,
-            SharedPreferences.OnSharedPreferenceChangeListener listener) {
+    public static void registerOnSharedPreferenceChangeListener(String key, SharedPreferences.OnSharedPreferenceChangeListener listener) {
         getSecurePreferences().registerOnSharedPreferenceChangeListener(key, listener);
     }
-
 
     /**
      * Unregisters a previous callback.
